gl-aip/examples/programmatic-tool-calling/pyproject.toml
[project]
name = "programmatic-tool-calling"
version = "0.1.0"
description = "Programmatic Tool Calling Example"
requires-python = ">=3.11,<3.13"
dependencies = [
    "glaip-sdk[local]>=0.8",
    "python-dotenv>=1.0.0",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
pythonpath = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
